[KI-337] application property for save mode added (append or overwrite)

+++ spark-importer/src/main/java/de/viadee/ki/sparkimporter/runner/CSVImportAndProcessingRunner.java
@@ -17,6 +17,7 @@ import de.viadee.ki.sparkimporter.util.SparkImporterVariables;
 import org.apache.commons.io.FileUtils;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
+import org.apache.spark.sql.SaveMode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -48,6 +49,7 @@ public class CSVImportAndProcessingRunner extends SparkRunner {
         SparkImporterVariables.setTargetFolder(ARGS.getFileDestination());
         SparkImporterVariables.setDevTypeCastCheckEnabled(ARGS.isDevTypeCastCheckEnabled());
         SparkImporterVariables.setRevCountEnabled(ARGS.isRevisionCount());
+        SparkImporterVariables.setSaveMode(ARGS.getSaveMode() == SparkImporterVariables.SAVE_MODE_APPEND ? SaveMode.Append : SaveMode.Overwrite);
         SparkImporterUtils.setWorkingDirectory(ARGS.getWorkingDirectory());
         SparkImporterLogger.setLogDirectory(ARGS.getLogDirectory());
 
@@ -56,7 +58,9 @@ public class CSVImportAndProcessingRunner extends SparkRunner {
         PreprocessingRunner.writeStepResultsIntoFile = ARGS.isWriteStepResultsToCSV();
 
         // Delete destination files, required to avoid exception during runtime
-        FileUtils.deleteQuietly(new File(ARGS.getFileDestination()));
+        if(SparkImporterVariables.getSaveMode().equals(SaveMode.Overwrite)) {
+            FileUtils.deleteQuietly(new File(ARGS.getFileDestination()));
+        }
 
         // register our own aggregation function
         sparkSession.udf().register("AllButEmptyString", new AllButEmptyStringAggregationFunction());
